Week_2/HFImageApp/requirements.txt
streamlit
huggingface_hub
transformers
diffusers
python-dotenv
torch
Pillow
accelerate